requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cssselect==1.3.0
curl_cffi==0.12.0
cycler==0.12.1
dnspython==2.7.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.59.0
frozendict==2.4.6
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jieba3k==0.35.1
joblib==1.5.1
kiwisolver==1.4.7
lxml==6.0.0
lxml_html_clean==0.4.2
markdown==3.7
matplotlib==3.9.4
multitasking==0.0.12
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yfinance==0.2.65
zipp==3.23.0
xhtml2pdf==0.2.15
